refactor: relocate utility classes and enhance static analysis configuration
- Move utility classes (`SplashUtils`, `TransitionUtils`, `ToastUtils`, etc.) from `ui.util` to `domain` package
- Add `URLUtils.kt` to centralize external link handling and string encoding logic
- Replace local `openURL` implementations in `AboutActivity` and `CustomAboutActivity` with domain utility
- Update `detekt.yml` with new rules for coroutines, naming conventions, and exception handling
- Clean up unused imports and simplify code patterns in `App` and `AppPickerActivity`
- Update Android Lint configuration to remove `GradleDependency` suppression
