+++ app/src/main/java/de/lemke/oneuisample/App.kt
@@ -2,6 +2,8 @@ package de.lemke.oneuisample
 
 import android.app.Application
 import androidx.appcompat.app.AppCompatDelegate
+import androidx.appcompat.app.AppCompatDelegate.MODE_NIGHT_NO
+import androidx.appcompat.app.AppCompatDelegate.MODE_NIGHT_YES
 import dagger.hilt.android.HiltAndroidApp
 import de.lemke.oneuisample.domain.GetUserSettingsUseCase
 import javax.inject.Inject
@@ -22,9 +24,9 @@ class App : Application() {
             val userSettings = getUserSettings()
             if (!userSettings.autoDarkMode) {
                 if (userSettings.darkMode) {
-                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
+                    AppCompatDelegate.setDefaultNightMode(MODE_NIGHT_YES)
                 } else {
-                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
+                    AppCompatDelegate.setDefaultNightMode(MODE_NIGHT_NO)
                 }
             }
         }
