+++ app/src/main/java/de/lemke/oneuisample/ui/MainActivity.kt
@@ -16,14 +16,14 @@ import de.lemke.oneuisample.R
 import de.lemke.oneuisample.databinding.ActivityMainBinding
 import de.lemke.oneuisample.domain.CheckAppStartUseCase
 import de.lemke.oneuisample.domain.GetUserSettingsUseCase
+import de.lemke.oneuisample.domain.configureSplashScreen
+import de.lemke.oneuisample.domain.finishWithFade
+import de.lemke.oneuisample.domain.onNavigationSingleClick
+import de.lemke.oneuisample.domain.onboardIfNeeded
+import de.lemke.oneuisample.domain.overrideFadeOpenTransition
+import de.lemke.oneuisample.domain.setupHeaderAndNavRail
+import de.lemke.oneuisample.domain.suggestiveSnackBar
 import de.lemke.oneuisample.ui.fragments.FragmentBottomSheet
-import de.lemke.oneuisample.ui.util.configureSplashScreen
-import de.lemke.oneuisample.ui.util.finishWithFade
-import de.lemke.oneuisample.ui.util.onNavigationSingleClick
-import de.lemke.oneuisample.ui.util.onboardIfNeeded
-import de.lemke.oneuisample.ui.util.overrideFadeOpenTransition
-import de.lemke.oneuisample.ui.util.setupHeaderAndNavRail
-import de.lemke.oneuisample.ui.util.suggestiveSnackBar
 import dev.oneuiproject.oneui.navigation.setupNavigation
 import javax.inject.Inject
 import kotlinx.coroutines.launch
