+++ app/src/main/java/de/lemke/oneuisample/ui/fragments/TabIcons.kt
@@ -33,10 +33,10 @@ import de.lemke.oneuisample.domain.GetUserSettingsUseCase
 import de.lemke.oneuisample.domain.ObserveIconListUseCase
 import de.lemke.oneuisample.domain.ObserveUserSettingsUseCase
 import de.lemke.oneuisample.domain.UpdateUserSettingsUseCase
+import de.lemke.oneuisample.domain.autoCleared
+import de.lemke.oneuisample.domain.suggestiveSnackBar
 import de.lemke.oneuisample.ui.util.IconAdapter
 import de.lemke.oneuisample.ui.util.IconAdapter.Icon
-import de.lemke.oneuisample.ui.util.autoCleared
-import de.lemke.oneuisample.ui.util.suggestiveSnackBar
 import dev.oneuiproject.oneui.delegates.AppBarAwareYTranslator
 import dev.oneuiproject.oneui.delegates.ViewYTranslator
 import dev.oneuiproject.oneui.ktx.dpToPx
