+++ config/detekt/detekt.yml
@@ -1,3 +1,25 @@
+config:
+  validation: true
+  warningsAsErrors: false
+
+processors:
+  active: true
+  exclude: [ ]
+
+console-reports:
+  active: true
+
+coroutines:
+  active: true
+  GlobalCoroutineUsage:
+    active: true
+  RedundantSuspendModifier:
+    active: true
+  SleepInsteadOfDelay:
+    active: true
+  SuspendFunWithFlowReturnType:
+    active: true
+
 complexity:
   active: true
   LongParameterList:
@@ -8,33 +30,54 @@ complexity:
     active: false
   ComplexCondition:
     allowedConditions: 4
-  LongMethod:
-    allowedLines: 70
 
 style:
   active: true
   MagicNumber:
     active: false # re-enable in Step 3 after extracting constants
-  WildcardImport:
-    active: true
+    ignorePropertyDeclaration: true
+    ignoreCompanionObjectPropertyDeclaration: true
   MaxLineLength:
     maxLineLength: 140
   ForbiddenComment:
     active: true
-    comments: ['TODO:', 'FIXME:']
+    comments: [ 'TODO:', 'FIXME:' ]
   UnusedPrivateFunction:
     active: true
   UnusedPrivateProperty:
     active: true
-  FunctionOnlyReturningConstant:
-    active: false
+
+naming:
+  FunctionNaming:
+    functionPattern: '[a-z][A-Za-z0-9]*'
+  TopLevelPropertyNaming:
+    constantPattern: '[A-Z][_A-Z0-9]*'
 
 empty-blocks:
   EmptyFunctionBlock:
     active: true
     ignoreOverridden: true
 
+exceptions:
+  active: true
+  TooGenericExceptionThrown:
+    active: true
+    exceptionNames:
+      - Error
+      - Exception
+      - RuntimeException
+      - Throwable
+  SwallowedException:
+    active: true
+    ignoredExceptionTypes:
+      - InterruptedException
+      - MalformedURLException
+      - NumberFormatException
+      - ParseException
+
 potential-bugs:
   active: true
-  ImplicitDefaultLocale: { active: true }
-  UnsafeCallOnNullableType: { active: true }
+  ImplicitDefaultLocale:
+    active: true
+  UnsafeCallOnNullableType:
+    active: true
